Create Elementary Particles
The scientists want to conduct a reaction to get two different subsegments of the same length. Denote this length k. The resulting pair of subsegments must be harmonious, i. e. for some i (1≤i≤k) it must be true that the types of particles on the i-th position are the same for these two subsegments. For example, the pair (1 7 3) and (4 7 8) is harmonious, as both subsegments have 7 on the second position. The pair (1 2 3) and (3 1 2) is not harmonious.
Hoặc tìm 2 số giống nhau có khoảng cách min trong 1 dãy số