Create All are same
Polycarp has an array of n (n is even) integers a1,a2,…,an. Polycarp conceived of a positive integer k. After that, Polycarp began performing the following operations on the array: take an index i (1≤i≤n) and reduce the number ai by k.

After Polycarp performed some (possibly zero) number of such operations, it turned out that all numbers in the array became the same. Find the maximum k at which such a situation is possible, or print −1 if such a number can be arbitrarily large.

Tutorial:  k can be arbitrarily large if and only if all numbers in the array are the same. In this case, we can choose any number k and subtract it from all the numbers, for example, exactly once.

Suppose we fix some k. Let qi be the number of subtractions of the number k from the number ai. In this case, all numbers will be equal if and only if, for any two numbers ai and aj from the array, ai−k⋅qi=aj−k⋅qj. Let qi0 be the minimum of qi. Then all numbers in the array become the same if for each index i we subtract from ai k not qi, but qi−qi0 times. Then we will never subtract k from the element ai0. This means that there is always an element in the array from which we can never subtract k. This element is the minimum on the array. Then from ai we will subtract k exactly ai−ai0k times.

Thus, with the current k, it is possible to make all elements equal if and only if for all elements ai the value ai−ai0 (where ai0 is the minimum on the array) is divisible by k. So, the maximum k is the greatest common divisor of all values of ai−ai0, i=1,n¯¯¯¯¯¯¯¯.