Create Update Files
Berland State University has received a new update for the operating system. Initially it is installed only on the 1-st computer.

Update files should be copied to all n computers. The computers are not connected to the internet, so the only way to transfer update files from one computer to another is to copy them using a patch cable (a cable connecting two computers directly). Only one patch cable can be connected to a computer at a time. Thus, from any computer where the update files are installed, they can be copied to some other computer in exactly one hour.

Your task is to find the minimum number of hours required to copy the update files to all n computers if there are only k patch cables in Berland State University.

Tutorial: Let cur be the current number of computers with the update already installed (initially it is 1). Then, in 1 hour, we can increase cur by min(cur,k). From here we can see that the value of cur will double for the first few hours, and then, when it becomes greater than k, it will begin to increase by exactly k.

The process when the number of computers doubles can be modeled using a loop, because the number of doublings does not exceed logn. And after that, we have to increase the answer by ⌈(n−cur)/k⌉ to take the number of additions of k into account.

Note that computing ⌈n−curk⌉ should be done without using fractional data types; to calculate ⌈x/y⌉ in integers, you should divide x+y−1 by y using integer division (this will work provided that both x and y are non-negative, and y≠0). If you use real numbers, this may cause precision issues.

Solution (Neon)